Updated README & example for additional User data.
Removed the "SecureUserFunc" reference from the README and replaced with the (correct)
example using auth.SecureUser and passing auth.User when needing access to additional                                                                       User parameters.

Also updated the Google example to reflect this change and included a /private/user
route that shows the additional user data in action.

The Google example README was fixed to correct a typo with the "access_key" flag.